requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
boto3==1.35.14
botocore==1.35.14
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cohere==5.9.1
colorama==0.4.6
cowsay==6.1
distro==1.9.0
fastavro==1.9.7
filelock==3.15.4
Flask==3.0.3
fsspec==2024.9.0
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-speech==2.27.0
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
llvmlite==0.43.0
MarkupSafe==2.1.5
more-itertools==10.5.0
mpmath==1.3.0
mutagen==1.47.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
openai==1.44.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
packaging==24.1
parameterized==0.9.0
proto-plus==1.24.0
protobuf==5.28.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycryptodomex==3.20.0
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
python-dateutil==2.9.0.post0
pytube==15.0.0
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rsa==4.9
s3transfer==0.10.2
setuptools==74.1.2
six==1.16.0
sniffio==1.3.1
sympy==1.13.2
tiktoken==0.7.0
tokenizers==0.20.0
torch==2.4.1
tqdm==4.66.5
types-requests==2.32.0.20240907
typing_extensions==4.12.2
urllib3==2.2.1
websockets==13.0.1
Werkzeug==3.0.4
yt-dlp==2024.8.6
